+++ src/net-util/export/DispatchInfo.js
@@ -6,6 +6,7 @@ import { BaseConverter, Struct } from '@this/data-values';
 import { MustBe } from '@this/typey';
 
 import { IncomingRequest } from '#x/IncomingRequest';
+import { UriUtil } from '#x/UriUtil';
 
 
 /**
@@ -57,17 +58,15 @@ export class DispatchInfo {
   }
 
   /**
-   * @returns {string} {@link #base}, as a path string. It is always prefixed
-   * with a slash (`/`) and only ends with a slash if the final path component
-   * is empty.
+   * @returns {string} {@link #base}, as an absolute path string. (See
+   * {@link UriUtil#pathStringFrom}).
    *
    * **Note:** It is strongly recommended to _only_ use this property for
    * logging/debugging purposes. {@link #base} is a better choice when doing
    * calculations.
    */
   get baseString() {
-    // `true` == relative form.
-    return this.#base.toUriPathString();
+    return UriUtil.pathStringFrom(this.#base);
   }
 
   /**
@@ -79,16 +78,16 @@ export class DispatchInfo {
   }
 
   /**
-   * @returns {string} {@link #extra}, as a path string. It is always prefixed
-   * with a dot-slash (`./`) and only ends with a slash if the final path
-   * component is empty.
+   * @returns {string} {@link #extra}, as a relative path string. (See
+   * {@link UriUtil#pathStringFrom}).
    *
    * **Note:** It is strongly recommended to _only_ use this property for
-   * logging/debugging purposes. {@link #base} is a better choice when doing
+   * logging/debugging purposes. {@link #extra} is a better choice when doing
    * calculations.
    */
   get extraString() {
-    return this.#extra.toUriPathString();
+    // `true` == relative form.
+    return UriUtil.pathStringFrom(this.#extra, true);
   }
 
   /**