+++ src/net-util/export/IncomingRequest.js
@@ -5,8 +5,7 @@ import { IncomingMessage } from 'node:http';
 import { Http2ServerRequest } from 'node:http2';
 
 import { TreePathKey } from '@this/collections';
-import { FormatUtils } from '@this/loggy-intf';
-import { IntfLogger } from '@this/loggy-intf';
+import { FormatUtils, IntfLogger } from '@this/loggy-intf';
 import { MustBe } from '@this/typey';
 
 import { Cookies } from '#x/Cookies';